small fixes

+++ mpi/common/common.h
@@ -1,10 +1,10 @@
 #include <iostream>
 
-bool IsPrime(int number) {
+bool IsPrime(size_t number) {
   if (number == 2 || number == 3) return true;
   if (number % 2 == 0 || number % 3 == 0) return false;
 
-  int divisor = 6;
+  size_t divisor = 6;
   while (divisor * divisor - 2 * divisor + 1 <= number) {
     if (number % (divisor - 1) == 0) return false;
     if (number % (divisor + 1) == 0) return false;
@@ -14,13 +14,13 @@ bool IsPrime(int number) {
   return true;
 }
 
-int NextPrime(int a) {
+int NextPrime(size_t a) {
   while (!IsPrime(++a)) {
   }
   return a;
 }
 
-int NextPrime(int a, int n) {
+int NextPrime(size_t a, size_t n) {
   for (size_t i = 0; i < n; i++)
     while (!IsPrime(++a)) {
     }
@@ -40,7 +40,7 @@ struct PrimeLine {
     first_ = n;
     sum_ = first_;
 
-    for (int i = 1; i < k; i++) {
+    for (size_t i = 1; i < k; i++) {
       n = NextPrime(n);
       sum_ += n;
     }
